move handlers and parses services into separate folders
