+++ DataReader.Core/Abstractions/Services/Handlers/IUserHandlerService.cs
@@ -1,0 +1,7 @@
+﻿namespace DataReader.Core.Abstractions.Services.Handlers
+{
+    public interface IUserHandlerService
+    {
+        Task Sync(string json);
+    }
+}
