--- DataReader.Core/Abstractions/Services/IUserHandlerService.cs
@@ -1,7 +1,0 @@
-﻿namespace DataReader.Core.Abstractions.Services
-{
-  public interface IUserHandlerService
-  {
-    Task Sync(string json);
-  }
-}
