+++ DataReader.Core/Abstractions/Services/Parsers/IJsonParserService.cs
@@ -1,0 +1,11 @@
+﻿using CSharpFunctionalExtensions;
+using DataReader.Core.Contracts.Params;
+
+
+namespace DataReader.Core.Abstractions.Services.Parsers
+{
+    public interface IJsonParserService
+    {
+        Result<List<UsersDTOParam>?> ParseUser(string json);
+    }
+}
